requirements.txt
asgiref==3.5.0
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
